+++ src/components/util/util.test.js
@@ -425,11 +425,11 @@ describe('isAffiliationsEnabled', () => {
     expect(isAffiliationsEnabled({ type: USER_TYPES.PATRON })).toBeFalsy();
   });
 
-  it('should return \'false\' if a user type is \'dbc\'', () => {
+  it('should return \'false\' if a user type is \'dcb\'', () => {
     expect(isAffiliationsEnabled({ type: USER_TYPES.DCB })).toBeFalsy();
   });
 
-  it('should return \'true\' if a user type is other than \'patron\' and \'dbc\'', () => {
+  it('should return \'true\' if a user type is other than \'patron\' and \'dcb\'', () => {
     expect(isAffiliationsEnabled({ type: USER_TYPES.SHADOW })).toBeTruthy();
     expect(isAffiliationsEnabled({ type: USER_TYPES.STAFF })).toBeTruthy();
     expect(isAffiliationsEnabled({ type: USER_TYPES.SYSTEM })).toBeTruthy();